requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
pillow
svgwrite
ezdxf
requests